feat(rayjob): RayJob runner implementation
This commit adds a runner for RayJobs in a Kubernetes cluster
with the Kuberay operator installed.

It uses the custom image built by our tooling as the runtime
environment for both the ephemeral Ray cluster and the job
submission pods.

The container YAML spec has been extended to better configure
the user and ownership of files copied into the image.

Images built as part of our job submission logic are optionally
pushed to a remote image registry, so they can be pulled by
the Kubernetes cluster as necessary (i.e., for Kueue and
RayJob executions).