Use single K8sResourceOption type
as there is no conceptual difference within the two. As they may use
different resources (in the limit and request) we do make the parameters
optional. This also reflects the code logic which filters `None` values.

We now also handle LIMIT and REQUEST types the same. This commit leaves
a TODO that we have to decide if separate logic is warrented or not.
Until then we do not remove the logical branches from the code, but
right as of this commit both branches act the same.