Replace raw config `dicts` with `TypedDicts` (#18)
* Add Typing for custom ResourceOptions

by introducing TypedDicts for K8s, Docker, and Ray ResourceOptions.

their filling value can be `None`. Then, we remove the `None`s by
using the `remove_none_values` function.

We add generic typing to the `remove_none_value` function and coerce back to the input type for consisten type hints.

As the evaluation of the functional form is not deferred and using
future imports has no effect, we use Union instead of | (pipe) to be compatible with
Python 3.9

* Use single K8sResourceOption type

as there is no conceptual difference within the two. As they may use
different resources (in the limit and request) we do make the parameters
optional. This also reflects the code logic which filters `None` values.

We now also handle LIMIT and REQUEST types the same. This commit leaves
a TODO that we have to decide if separate logic is warranted or not.
Until then we do not remove the logical branches from the code, but
right as of this commit both branches act the same.