Move Oldest supported Python version to 3.10 (#21)
* Move oldest supported Python up to 3.10

in the pyproject.toml and the CI tests.

* Use TypeAlias instead of Union Type for Path

we had this implementation previously but changed it to be compatible
with Python 3.9. As we now move to >=3.10 we can use this more elegant
solution.

* Use slots for config dataclasses

We had this implementation earlier but removed it to support Python 3.9.
As we now move to Python 3.10 we can reintroduce them.

* Use Pipes instead of Union for union types

as we now move to Python 3.10 and pipes are easier to read.