Simplify smoke test examples resource consumption
such that it puts less strain on developer machines.
Also change the namespace from 'mlops' back to 'localhost' to
avoid manual namespace setup in a new cluster instance upon testing.

If more resources or a custom namespace is required, we can set it up
quickly when needed. Otherwise, for the default, this setup will be
more convenient.