fix(docker): Fix ordering of user creation and use in Dockerfile
Users are now created before any Python packages are installed.
Hence, all Python packages are installed locally for the user, not
system-wide (preventing accidental override of the system
interpreter as well).

The user's `~/.local/bin` folder is included in the PATH for the image,
so that installed Python scripts are found correctly.

Buildx cache folders for pip are mounted to the correct destination
(i.e., the user's `~/.cache/pip` folder).