Attach committer to job metadata
via the labels of the docker container. These labels are also added
to the annotations of k8s for Kueue and Ray.

The submitter can either be submitted via an optional CLI arg or,
when it is not set, we try to infer it first from the git author
and then from the author of the last commit.
If both does not work we set the submitter to "Unkonwn"

We use an underscore "_submitter" to avoid namespace collisions in
the labels.