feat(client): Add backend API client based on OpenAPI-generator
The code will not be automatically generated in the future,
rather the underlying models come from the client package,
with just the top-level API endpoint models being defined in
the API client package.

This allows us to have a single source of truth when it comes
to the data model of the `Job` and its options.