Add shell script for automatic OpenAPI client regeneration
Needs a running FastAPI backend on localhost port 8000.

Downloads its OpenAPI spec into a unique temporary file, runs
openapi-gen in a docker container on it, and moves the generated
client code (and only the source code) into the `client` subpackage.

This includes the docs and test directories, which contain autogen'd
info that we might not care about.