feat(backend): Add Dockerfile and Helm chart
The Dockerfile implements a multi-stage build to create a container that
runs the FastAPI application.

The Helm chart is used to deploy the application to a Kubernetes cluster.
It includes the necessary RBAC resources to grant access to the relevant
resources in the cluster (mainly Kueue resources and Jobs/RayJobs).

Issue: #68