Enable multi-pod log collection in both ex-post and streaming modes
The streaming case is really difficult, since the previous "stream" is
a (synchronous) generator around an urllib response body.

In the multi-pod case, we just stream these responses (from the k8s API)
line-by-line. It seems that because of this synchronicity, in the worst
case, we wait for the slowest pod to log again, which can potentially
stall log output from all pods indefinitely. It is unclear how to resolve
this without going to a fully asynchronous approach.

In the static case, we simply fetch all logs and concatenate them into
a big string.