Refactor build context resolution to Job class
as the prior resolution in the `ImageOptions` has the drawback that
the `Job` is not yet defined. This made the relative resolution of
the build context dependent on brittle runtime inspection.

Now we resolve the build context heuristically to the root of the job
project (via e.g. the presetce of a pyproject.toml).

Relative paths to the dockerfile or .yaml spec must be relative to
this resolved root.