Enable multi-pod log collection in both ex-post and streaming modes (#88)
* Enable multi-pod log collection in both ex-post and streaming modes

The streaming case is really difficult, since the previous "stream" is
a (synchronous) generator around an urllib response body.

In the multi-pod case, we just stream these responses (from the k8s API)
line-by-line. It seems that because of this synchronicity, in the worst
case, we wait for the slowest pod to log again, which can potentially
stall log output from all pods indefinitely. It is unclear how to resolve
this without going to a fully asynchronous approach.

In the static case, we simply fetch all logs and concatenate them into
a big string.

* hack: Introduce minimal workload mock to pass pod number checks

I didn't find a more graceful solution in an economical amount of time,
so I went with a class-level minimal workload mock that codifies our
assumption of "at least one pod is found in a logs streaming scenario".

The previous non-value mock had the issue that it did not pass length
checks and iteration over the pod list, so the mocked log streaming APIs
were never called.

* Add async reading with a task queue to stream multipod logs efficiently

This should fix the waiting collective problem introduced by synchronous
reading in the previous solution.

* fix(backend): Fix test regression for log streaming

---------

Co-authored-by: Adrian Rumpold <a.rumpold@gmail.com>