Handle relative (to job) definition of spec when executed in other directory (#84)
* Error Handling (client) Change File not found Error Msg

because the `is_file` check results in a unpleasantly trailing `False`.

* WIP -- feat(client) Take spec path relative to file where job defined

by getting the location of the job definition at runtime and
"subtracting" the wrong initialisation from Path (relative to
execution directory) to get a relative path to the spec.

MISSING: Handle absolute paths to spec.

* (wip) Raw string manipulation for spec path recovery

* feat(client) Resolve relative spec path relative to job

not relative to cwd. Absolute paths are handled as such.

Also set build context as parent of spec.

* fixup(tests) Set spec path relative to job definition

as we implemented path resolution relative to job.

* chore(client) Move examples into dedicated example folder

now that we support resolution of the spec path relative to the
job definiton

* Refactor build context resolution to Job class

as the prior resolution in the `ImageOptions` has the drawback that
the `Job` is not yet defined. This made the relative resolution of
the build context dependent on brittle runtime inspection.

Now we resolve the build context heuristically to the root of the job
project (via e.g. the presetce of a pyproject.toml).

Relative paths to the dockerfile or .yaml spec must be relative to
this resolved root.

* Enforce Pathlib with Ruff

and fix resulting linting errors.

* Small cleanups

* Exclude openapi auto generated code from linting

* chore(client) Reviewers comments

* fix(client): Fix path resolution errors in workflow image

* Resolve build ctx from entrypoint abs-path and file relative ctx

---------

Co-authored-by: Adrian Rumpold <a.rumpold@gmail.com>