feat(backend): Support for image pull policies
This commit adds support for pull policies in the runners.

The job submission endpoint passes the pull policy to the runner, by
default the policy is set to `Always` to prevent issues with stale
images.

The e2e test has been updated to pass `IfNotPresent` as the pull policy,
since the job image for the test suite is built locally and cannot be
pulled from a registry.